discover/grub2: Fix default option check
A default menu entry can be chosen using any of the following attributes:
index, title or id (if the entry was defined with the --id option).

But Petitboot doesn't honor this correctly and only compares the default
with the menu entry title if the entry doesn't have an id defined. This
is wrong since an index or title can be used even if an id was defined.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>